# SpeerNotes/bin/Release/net8.0/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "SpeerNotes, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
    "/api/auth/signup": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/SignUpRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/SignUpResponse"
            }
          }
        }
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/TokenResponse"
            }
          }
        }
      }
    },
    "/api/notes": {
      "get": {
        "tags": [
          "Notes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetAllNotesResponse"
            }
          }
        }
      },
      "post": {
        "tags": [
          "Notes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/CreateNoteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/CreateNoteResponse"
            }
          }
        }
      },
      "put": {
        "tags": [
          "Notes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/UpdateNoteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponseModel"
            }
          }
        }
      }
    },
    "/api/notes/{id}": {
      "get": {
        "tags": [
          "Notes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/GetNoteResponse"
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Notes"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponseModel"
            }
          }
        }
      }
    },
    "/api/notes/{id}/share": {
      "post": {
        "tags": [
          "Notes"
        ],
        "consumes": [
          "application/json",
          "text/json",
          "application/*+json"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "schema": {
              "$ref": "#/definitions/ShareNoteRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponseModel"
            }
          }
        }
      }
    },
    "/api/search": {
      "get": {
        "tags": [
          "Search"
        ],
        "produces": [
          "text/plain",
          "application/json",
          "text/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/BaseResponseModel"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "BaseResponseModel": {
      "type": "object",
      "properties": {
        "successful": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationError"
          }
        }
      },
      "additionalProperties": false
    },
    "CreateNoteRequest": {
      "required": [
        "details",
        "title"
      ],
      "type": "object",
      "properties": {
        "title": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "details": {
          "maxLength": 500,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "CreateNoteResponse": {
      "type": "object",
      "properties": {
        "successful": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationError"
          }
        },
        "noteId": {
          "format": "int64",
          "type": "integer"
        }
      },
      "additionalProperties": false
    },
    "GetAllNotesResponse": {
      "type": "object",
      "properties": {
        "successful": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationError"
          }
        },
        "data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/Note"
          }
        }
      },
      "additionalProperties": false
    },
    "GetNoteResponse": {
      "type": "object",
      "properties": {
        "successful": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationError"
          }
        },
        "data": {
          "$ref": "#/definitions/Note"
        }
      },
      "additionalProperties": false
    },
    "LoginRequest": {
      "required": [
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "userName": {
          "maxLength": 50,
          "minLength": 6,
          "type": "string"
        },
        "password": {
          "maxLength": 50,
          "minLength": 6,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "Note": {
      "required": [
        "createdBy",
        "createdOn",
        "details",
        "title"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "title": {
          "type": "string"
        },
        "details": {
          "type": "string"
        },
        "createdBy": {
          "type": "string"
        },
        "createdOn": {
          "format": "date-time",
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ShareNoteRequest": {
      "required": [
        "shareWith"
      ],
      "type": "object",
      "properties": {
        "shareWith": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "SignUpRequest": {
      "required": [
        "confirmPassword",
        "password",
        "userName"
      ],
      "type": "object",
      "properties": {
        "userName": {
          "maxLength": 50,
          "minLength": 6,
          "pattern": "^[a-zA-Z][a-zA-Z0-9 -_\\|]+$",
          "type": "string"
        },
        "password": {
          "maxLength": 50,
          "minLength": 6,
          "type": "string"
        },
        "confirmPassword": {
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "SignUpResponse": {
      "type": "object",
      "properties": {
        "successful": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationError"
          }
        }
      },
      "additionalProperties": false
    },
    "TokenResponse": {
      "type": "object",
      "properties": {
        "successful": {
          "type": "boolean",
          "readOnly": true
        },
        "errors": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/ValidationError"
          }
        },
        "token": {
          "type": "string",
          "readOnly": true
        },
        "validTo": {
          "format": "date-time",
          "type": "string",
          "readOnly": true
        }
      },
      "additionalProperties": false
    },
    "UpdateNoteRequest": {
      "required": [
        "details",
        "title"
      ],
      "type": "object",
      "properties": {
        "id": {
          "format": "int64",
          "type": "integer"
        },
        "title": {
          "maxLength": 50,
          "minLength": 1,
          "type": "string"
        },
        "details": {
          "maxLength": 500,
          "minLength": 1,
          "type": "string"
        }
      },
      "additionalProperties": false
    },
    "ValidationError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        },
        "description": {
          "type": "string"
        }
      },
      "additionalProperties": false
    }
  },
  "securityDefinitions": {
    "Bearer": { }
  },
  "security": [
    {
      "Bearer": [ ]
    }
  ]
}